+++ apps/client-e2e/src/tests/notebooks/run-problem-notebooks.ts
@@ -1,7 +1,10 @@
 import { GetExtensionPath } from '@idl/shared';
 
 import { RunnerFunction } from '../runner.interface';
-import { RunNotebookAndCompareCells } from './helpers/run-notebook-and-compare-cells';
+import {
+  DEFAULT_RUNNER_TIMEOUT,
+  RunNotebookAndCompareCells,
+} from './helpers/run-notebook-and-compare-cells';
 
 /**
  * Function that verifies that we can do basic debugging of IDL sessions
@@ -35,7 +38,8 @@ export const RunProblemNotebooks: RunnerFunction = async (init) => {
         mimeTypes: [],
       },
     ],
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 
   /**
@@ -51,7 +55,8 @@ export const RunProblemNotebooks: RunnerFunction = async (init) => {
         mimeTypes: [],
       },
     ],
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 
   /**
@@ -73,7 +78,8 @@ export const RunProblemNotebooks: RunnerFunction = async (init) => {
         mimeTypes: [],
       },
     ],
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 
   /**
@@ -96,6 +102,7 @@ export const RunProblemNotebooks: RunnerFunction = async (init) => {
         mimeTypes: [],
       },
     ],
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 };