+++ apps/client-e2e/src/tests/notebooks/run-test-envi-map-notebook.ts
@@ -3,7 +3,10 @@ import { GetExtensionPath } from '@idl/shared';
 
 import { RunnerFunction } from '../runner.interface';
 import { ICompareCellOutputs } from './helpers/compare-cells.interface';
-import { RunNotebookAndCompareCells } from './helpers/run-notebook-and-compare-cells';
+import {
+  DEFAULT_RUNNER_TIMEOUT,
+  RunNotebookAndCompareCells,
+} from './helpers/run-notebook-and-compare-cells';
 
 /**
  * Types of outputs from cells that we expect to have
@@ -52,7 +55,7 @@ export const CELL_OUTPUT: ICompareCellOutputs[] = [
   {
     idx: 8,
     success: true,
-    mimeTypes: [],
+    mimeTypes: [IDL_NOTEBOOK_MIME_TYPE],
   },
 ];
 
@@ -65,6 +68,7 @@ export const RunTestENVIMapNotebook: RunnerFunction = async (init) => {
       'idl/test/client-e2e/notebooks/test-notebook-envi-maps.idlnb'
     ),
     CELL_OUTPUT,
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 };