+++ apps/client-e2e/src/tests/notebooks/run-test-envi-notebook.ts
@@ -3,7 +3,10 @@ import { GetExtensionPath } from '@idl/shared';
 
 import { RunnerFunction } from '../runner.interface';
 import { ICompareCellOutputs } from './helpers/compare-cells.interface';
-import { RunNotebookAndCompareCells } from './helpers/run-notebook-and-compare-cells';
+import {
+  DEFAULT_RUNNER_TIMEOUT,
+  RunNotebookAndCompareCells,
+} from './helpers/run-notebook-and-compare-cells';
 
 /**
  * Types of outputs from cells that we expect to have
@@ -83,6 +86,7 @@ export const RunTestENVINotebook: RunnerFunction = async (init) => {
   await RunNotebookAndCompareCells(
     GetExtensionPath('idl/test/client-e2e/notebooks/envi-test-notebook.idlnb'),
     CELL_OUTPUT,
-    init.notebooks.controller
+    init.notebooks.controller,
+    DEFAULT_RUNNER_TIMEOUT
   );
 };