+++ libs/parsing/syntax-tree/src/lib/remove-scope-detail-and-reset-token-cache.ts
@@ -2,7 +2,6 @@ import { CancellationToken } from '@idl/cancellation-tokens';
 
 import { IParsed } from './build-syntax-tree.interface';
 import { TreeRecurserBasic } from './recursion-and-callbacks/tree-recurser-basic';
-import { ResetTokenCache } from './reset-token-cache';
 
 /**
  * Removes scope detail and resets our token cache
@@ -11,25 +10,18 @@ export function RemoveScopeDetailAndResetTokenCache(
   parsed: IParsed,
   cancel: CancellationToken
 ): void {
-  switch (true) {
-    case parsed.hasDetail:
-      TreeRecurserBasic(parsed.tree, cancel, {
-        onBasicToken: (token) => {
-          delete token.scopeTokens;
-          delete token.accessTokens;
-          token.cache = {};
-        },
-        onBranchToken: (token) => {
-          delete token.scopeTokens;
-          delete token.accessTokens;
-          token.cache = {};
-        },
-      });
-      parsed.hasCache = false;
-      parsed.hasDetail = false;
-      break;
-    default:
-      ResetTokenCache(parsed, cancel);
-      break;
-  }
+  TreeRecurserBasic(parsed.tree, cancel, {
+    onBasicToken: (token) => {
+      delete token.scopeTokens;
+      delete token.accessTokens;
+      token.cache = {};
+    },
+    onBranchToken: (token) => {
+      delete token.scopeTokens;
+      delete token.accessTokens;
+      token.cache = {};
+    },
+  });
+  parsed.hasCache = false;
+  parsed.hasDetail = false;
 }