+++ libs/shared/src/lib/version.interface.ts
@@ -1,4 +1,4 @@
 /**
  * Version of the extension
  */
-export const VERSION = '3.2.3';
+export const VERSION = '3.2.4';