+++ CHANGELOG.md
@@ -42,15 +42,31 @@ Here are some of the features that notebooks bring:
 
   - After each cell is executed we issue a `retall` command to make sure that we are at the top-level and not stopped in a weird state
 
+## 3.2.4 October 2023
+
+Fixed an issue where improper versions of node.js were used for the language server startup. This caused failures in the language server starting which would prevent any of the feature goodness of the language server from being accessible.
+
+Added in a sidebar entry for notebooks
+
+Updated cache logic to fix a potential circular JSON error when adding items to our cache
+
+Renamed "Additional Actions" sidebar to "Quick Access" and updated the icon used
+
+Resolve an edge case when running notebooks where a false-failure would be detected that halts notebook execution.
+
+Fixed an issue when launching IDL on Mac where ENVI crashes because "LANG" is missing from the environment
+
 ## 3.2.3 September 2023
 
 Re-work cancellation to handle cases where the worker threads were too busy to get the messages that work needs to be stopped. If you now use a very large file (take "slicer3.pro" from the IDL distribution) events are interrupted as expected.
 
-This also includes a fix where flase error messages would show up in the UI for work that was canceled.
+This also includes a fix where false error messages would show up in the UI for work that was canceled.
 
 Update most dependencies with changes from first re-release of the extension
 
 Fixed a bug where the "strictarrsubs" compile option was missing from the list of valid compile options
+
+Fixed a bug with the code sidebar icon not showing up in light themes
 
 ## 3.2.2 September 2023